vaquita and d3 as separate dep libs, restore LA page as standard import
